requirements.txt
flask==3.0.3
flask-cors==5.0.0
google-generativeai==0.8.5
python-dotenv==1.1.1
